refactor(package-manager): extract inline test modules into dedicated files
Move two inline `#[cfg(test)] mod tests { ... }` blocks into external
`tests.rs` files following CODE_STYLE_GUIDE.md: `link_file` (402 lines)
and `install` (396 lines). No behavioral change.

Part of #358.
